@@ -13,13 +13,13 @@
         "google/apiclient": "^2.17",
         "guzzlehttp/guzzle": "^7.0.1",
         "intervention/image": "^2.6",
-        "laravel/framework": "^9.19",
+        "laravel/framework": "^10.0",
         "laravel/jetstream": "^3.0",
         "laravel/sanctum": "^3.2",
         "laravel/tinker": "^2.8",
         "livewire/livewire": "^2.11",
         "maatwebsite/excel": "^3.1",
-        "pusher/pusher-php-server": "^5.0",
+        "pusher/pusher-php-server": "^7.0",
         "spatie/laravel-permission": "^5.5",
         "symfony/http-client": "^7.3",
         "symfony/mailgun-mailer": "^7.1",
@@ -27,7 +27,7 @@
     },
     "require-dev": {
         "barryvdh/laravel-debugbar": "^3.6",
-        "spatie/laravel-ignition": "^1.0",
+        "spatie/laravel-ignition": "^2.0",
         "fakerphp/faker": "^1.9.1",
         "laravel/sail": "^1.0.1",
         "mockery/mockery": "^1.4.2",
